+++ src/components/playlist/SongList/SongBackground.tsx
@@ -1,13 +1,19 @@
 import React from 'react';
 import { StyleSheet, ImageBackground } from 'react-native';
+
 import { i0hdslbHTTPResolve } from '@utils/Utils';
+import { useNoxSetting } from '@stores/useApp';
 
 interface Props {
   song: NoxMedia.Song;
-  current?: boolean;
   children: React.JSX.Element;
 }
-const SongBackground = ({ song, current, children }: Props) => {
+const SongBackground = ({ song, children }: Props) => {
+  const currentPlayingId = useNoxSetting(state => state.currentPlayingId);
+  const playerSetting = useNoxSetting(state => state.playerSetting);
+  const current =
+    playerSetting.trackCoverArtCard && song.id === currentPlayingId;
+
   return current ? (
     <ImageBackground
       source={{ uri: i0hdslbHTTPResolve(song.cover) }}
